Расставляет принудительные переносы в блоках кода
Fix: #842
