+++ src/styles/blocks/code-fix.css
@@ -1,4 +1,6 @@
 /* https://github.com/doka-guide/platform/issues/838 */
 .code-fix {
   line-height: 0;
+  overflow-wrap: break-word;
+  word-break: break-word;
 }
